+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Actors can now be stopped by calling stop on an ActorRef
   - this sends a "StopMessage" to the Actor
   - all messages queued until the stop message is handled in the actor will still be executed
+- fix ActorAddress generation used for Wakeup Calls  
 
 # 0.0.4